payment-backend/payments/requirements.txt
asgiref==3.7.2
attrs==23.2.0
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
crispy-bootstrap4==2024.1
crispy-forms-bootstrap2==2024.1
cryptography==42.0.5
defusedxml==0.7.1
distlib==0.3.8
Django
django-allauth==0.61.1
django-crispy-forms==2.1
djangorestframework==3.14.0
dnspython==2.3.0
drf-spectacular==0.27.1
filelock==3.13.1
geographiclib==2.0
idna==3.6
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
oauthlib==3.2.2
paho-mqtt==1.6.1
platformdirs==4.2.0
psycopg2==2.9.9
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-etcd==0.4.5
python3-openid==3.2.0
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
requests-oauthlib==2.0.0
rpds-py==0.18.0
sqlparse==0.4.4
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.0.2
virtualenv==20.25.1
